toolkit/components/passwordmgr/test/test_bug_242956.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: 242956
<p id="display"></p>
<div id="content" style="display: none">
  <!-- pword is not a type=password input -->
  <form id="form1" action="formtest.js">
    <input  type="text" name="uname">
    <input  type="text" name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- uname is not a type=text input -->
  <form id="form2" action="formtest.js">
    <input  type="password" name="uname">
    <input  type="password" name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- two "pword" inputs, (text + password) -->
  <form id="form3" action="formtest.js">
    <input  type="text"       name="uname">
    <input  type="text"       name="pword">
    <input  type="password"   name="qword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- same thing, different order -->
  <form id="form4" action="formtest.js">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <input  type="text"       name="qword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- uname is not a type=text input (try a checkbox just for variety) -->
  <form id="form5" action="formtest.js">
    <input  type="checkbox" name="uname" value="">
    <input  type="password" name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- pword is not a type=password input (try a checkbox just for variety) -->
  <form id="form6" action="formtest.js">
    <input  type="text"     name="uname">
    <input  type="checkbox" name="pword" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <!-- pword is not a type=password input -->
  <form id="form7" action="formtest.js">
    <input  type="text" name="uname" value="testuser">
    <input  type="text" name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>


</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: 242956 (Stored password is inserted into a
    readable text input on a second page) **/
commonInit();

// Make sure that pwmgr only puts passwords into type=password <input>s.
// Might as well test the converse, too (username in password field).

function startTest() {
  var form, input;

  is($_(1, "uname").value, "", "Checking for unfilled username 1");
  is($_(1, "pword").value, "", "Checking for unfilled password 1");

  is($_(2, "uname").value, "testpass", "Checking for password not username 2");
  is($_(2, "pword").value, "", "Checking for unfilled password 2");

  is($_(3, "uname").value, "", "Checking for unfilled username 3");
  is($_(3, "pword").value, "testuser", "Checking for unfilled password 3");
  is($_(3, "qword").value, "testpass", "Checking for unfilled qassword 3");

  is($_(4, "uname").value, "testuser", "Checking for password not username 4");
  is($_(4, "pword").value, "testpass", "Checking for unfilled password 4");
  is($_(4, "qword").value, "", "Checking for unfilled qassword 4");

  is($_(5, "uname").value, "", "Checking for unfilled username 5");
  is($_(5, "pword").value, "testpass", "Checking for filled password 5");

  is($_(6, "uname").value, "", "Checking for unfilled username 6");
  is($_(6, "pword").value, "", "Checking for unfilled password 6");

  is($_(7, "uname").value, "testuser", "Checking for unmodified username 7");
  is($_(7, "pword").value, "",         "Checking for unfilled password 7");



  SimpleTest.finish();
}

window.addEventListener("runTests", startTest);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
